requirements.txt
# =============================================================================
# NEEDFUL – used by the application at runtime or tests
# =============================================================================

# Web framework & server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Validation & serialization
pydantic==2.12.5
python-multipart==0.0.6

# Auth (JWT + Google OAuth)
python-jose==3.5.0
google-auth>=2.23.0

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# AI (Cohere, OCR, image processing)
cohere==5.20.1
pytesseract==0.3.13
Pillow==12.1.1
opencv-python==4.13.0.92

# Background tasks
celery[redis]==5.4.0
redis==5.2.1

# SSE (notifications streaming)
sse-starlette==2.1.3

# Testing
pytest==8.3.4


# =============================================================================
# REDUNDANT – not imported or used in this codebase (kept for reference)
# =============================================================================

# Auth: app uses Google OAuth + JWT only; no local password hashing
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Dev tooling (formatting, linting)
black==24.4.2
isort==5.13.2

# HTTP client – not imported; TestClient uses requests
httpx==0.27.2

# DB: vector extension created in SQL but no pgvector/vector columns in models
pgvector==0.3.2

# Document libs – not imported anywhere
pypdf==6.7.0
pymupdf==1.26.7
python-docx==1.2.0

# CLI output – not imported
rich==13.9.0
